feat: 프로필 완성 페이지 디자인 및 구조 변경
