+++ app/newcomer/components/PeriodStep.tsx
@@ -1,44 +1,27 @@
 import { useState } from "react";
-import { useController, useForm } from "react-hook-form";
-import { BottomButton } from "@/app/components/BottomButton";
+import { useController, useFormContext } from "react-hook-form";
 import { DatePicker } from "@/components/DatePicker";
 import { Dimmed } from "@/components/Dimmed";
 import { Divider } from "@/components/Divider/Divider";
 import { Period } from "@/types/mentor";
+import { cn } from "@/utils/cn";
 import { toYYYYMMDD } from "@/utils/dateUtils";
-import { type ProfileForm, useProfileFormStore } from "../stores";
+import { ProfileForm } from "../stores";
 
-interface PeriodStepProps {
-  onClickNextStep: () => void;
-}
-
-export const PeriodStep = ({ onClickNextStep }: PeriodStepProps) => {
+export const PeriodStep = () => {
   const [periodType, setPeriodType] = useState<keyof Period>();
-  const { period, setProfileData } = useProfileFormStore();
 
-  const {
-    control,
-    handleSubmit,
-    formState: { isValid },
-  } = useForm<Pick<ProfileForm, "period">>();
+  const { control } = useFormContext<Pick<ProfileForm, "period">>();
 
   const TODAY = toYYYYMMDD(new Date());
   const { field: startDate } = useController({
     control,
     name: "period.startDate",
-    defaultValue: period?.startDate || TODAY,
-    rules: {
-      required: true,
-    },
   });
 
   const { field: endDate } = useController({
     control,
     name: "period.endDate",
-    defaultValue: period?.endDate || "",
-    rules: {
-      required: true,
-    },
   });
 
   const handleChangeDate = (YYYYMMDD: string) => {
@@ -50,36 +33,31 @@ export const PeriodStep = ({ onClickNextStep }: PeriodStepProps) => {
     setPeriodType(undefined);
   };
 
-  const handleClickNextStep = ({ period }: Pick<ProfileForm, "period">) => {
-    setProfileData({ period });
-    onClickNextStep();
-  };
-
   return (
-    <form onSubmit={handleSubmit(handleClickNextStep)}>
-      <div className="headline-1 mb-9">
-        어느 기간 내에
-        <br />
-        진행할 예정이신가요?
-      </div>
+    <>
       <div className="body-1 mb-2 text-gray-500">이때부터 시작할 예정이에요 (시작)</div>
-      <button className="headline-3" type="button" onClick={() => setPeriodType("startDate")}>
+      <button
+        className={cn("headline-3", !startDate.value && "text-gray-400")}
+        type="button"
+        onClick={() => setPeriodType("startDate")}
+      >
         {startDate.value || TODAY}
       </button>
-      <Divider className="my-6" />
+      <Divider className="my-[16px]" />
       <div className="body-1 mb-2 text-gray-500">이때까지 가능해요 (완료)</div>
-      <button className="headline-3" type="button" onClick={() => setPeriodType("endDate")}>
+      <button
+        className={cn("headline-3", !endDate.value && "text-gray-400")}
+        type="button"
+        onClick={() => setPeriodType("endDate")}
+      >
         {endDate.value || TODAY}
       </button>
-      <BottomButton type="submit" disabled={!isValid}>
-        다음
-      </BottomButton>
       {periodType && (
         <>
           <Dimmed onClick={() => setPeriodType(undefined)} />
           <DatePicker onChangeDate={handleChangeDate} />
         </>
       )}
-    </form>
+    </>
   );
 };
